refactor(database): refactoring: move pws database class to common module
