+++ app/src/main/kotlin/io/github/alelk/pws/android/app/di/AppModule.kt
@@ -1,7 +1,7 @@
 package io.github.alelk.pws.android.app.di
 
 import android.content.Context
-import io.github.alelk.pws.database.DatabaseProvider
+import io.github.alelk.pws.database.PwsDatabaseProvider
 import io.github.alelk.pws.database.PwsDatabase
 import dagger.Module
 import dagger.Provides
@@ -16,5 +16,5 @@ object AppModule {
 
   @Provides
   @Singleton
-  fun provideDatabase(@ApplicationContext context: Context): PwsDatabase = DatabaseProvider.getDatabase(context)
+  fun provideDatabase(@ApplicationContext context: Context): PwsDatabase = PwsDatabaseProvider.getDatabase(context)
 }
