+++ database/src/commonMain/kotlin/io/github/alelk/pws/database/entity/FavoriteEntity.kt
@@ -5,7 +5,11 @@ import androidx.room.*
 @Entity(
   tableName = "favorites",
   foreignKeys = [
-    ForeignKey(entity = SongNumberEntity::class, parentColumns = ["_id"], childColumns = ["psalmnumberid"], onDelete = ForeignKey.CASCADE)
+    ForeignKey(
+      entity = SongNumberEntity::class,
+      parentColumns = ["_id"], childColumns = ["psalmnumberid"],
+      onDelete = ForeignKey.CASCADE, onUpdate = ForeignKey.NO_ACTION
+    )
   ],
   indices = [
     Index(name = "idx_favorites_psalmnumberid", value = ["psalmnumberid"]),
@@ -15,5 +19,5 @@ import androidx.room.*
 data class FavoriteEntity(
   @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "_id") val id: Long = 0,
   @ColumnInfo(name = "position") val position: Int,
-  @ColumnInfo(name = "psalmnumberid", index = true) val songNumberId: Long
+  @ColumnInfo(name = "psalmnumberid") val songNumberId: Long
 )
